ICUnover socket UupdateCash parseInt 변경

+++ ICUnover/public/source/socket.js
@@ -19,7 +19,7 @@ socket.on('UserLogout', () => {
 
 socket.on('UpdateCash', (iCash) => {
 
-    const cCash = parseFloat(iCash);
+    const cCash = parseInt(iCash);
 
     // $('#MainLayoutCash').text(iCash);
     // $('#MainLayoutCoin').text(iCash);
